fix dir structure
