+++ src/procedure/output/uniqueDevice/led.ts
@@ -1,6 +1,6 @@
 import { outputPort } from '../outputPort'
 import type { SerialPort } from 'serialport'
-import { delay } from '../../../utils/delay'
+import { delay } from '../../utils/delay'
 
 export const attachLed = (port: SerialPort, pin: number) => {
   const led = outputPort(port)(pin)
