+++ src/procedure/pwm/uniqueDevice/passiveBuzzer.ts
