+++ src/procedure/pwm/uniqueDevice/vibrationSensor.ts
